file-to-md/file-to-md-app/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
PyMuPDF>=1.25.0
python-docx>=1.1.0
pytesseract>=0.3.13
Pillow>=11.0.0
beautifulsoup4>=4.13.0
markdownify>=0.14.0
requests>=2.32.0
